Fix float conversion warnings
including conversion and double-promotion

+++ KappaAnalysis/src/Producers/PUWeightProducer.cc
@@ -31,8 +31,8 @@ void PUWeightProducer::Produce(KappaEvent const& event, KappaProduct& product,
                      KappaSettings const& settings) const
 {
 	assert(event.m_genEventInfo != NULL);
-	
-	unsigned int puBin = (static_cast<double>(event.m_genEventInfo->nPUMean) * m_bins);
+
+	unsigned int puBin = static_cast<unsigned int>(static_cast<double>(event.m_genEventInfo->nPUMean) * m_bins);
 	if (puBin < m_pileupWeights.size())
 		product.m_weights["puWeight"] = m_pileupWeights.at(puBin);
 	else
